requirements.txt
falcon==1.4.1
requests==2.20.0
simplejson==3.13.2
grpcio==1.19.0
grpcio-tools==1.19.0
protobuf==3.7.1